New Anvil format; added GUI
Added support for .mca Anvil world format. This was a pain in the ass
for the following reasons:
- New data type in the NBT format.
- X and Z are no longer N-S and E-W. They are now W-E and N-S. This
makes more sense, but it was still a pain to change. (NOTE: The oblique
map has NOT YET been updated, so it'll still look fucked up.)

Also, adding in the PySide GUI launcher. Been in dev for a while but
wasn't in the repo.