Pasted selection now movable
Also, a neato dark background to show that there is a pasted
selection.

Pasted selection also gets redrawn with the rest of the map, for
instance when adding biomes.

Box tool now has to be explicitly selected to work.

Added whitelist to a world method that was missing the option.

Probably some other stuff too.