Pasted selection pixmaps are now re-rendered in the new tab
Previously, a pasted selection would just take the pixmaps from the old
tab, which would mismatch if the new tab was in a different view mode
(biome overlay, etc). The paste method has been taken out of the tab
class and put in the main class, and pixmaps are all re-rendered when
chunks are pasted into a new tab.