Saving seems to work...
Fixed an oversight in writing compound tags: forgot to write the final
0-byte at the end.

Also replaced the AnvilRegion subclass with a simple flag, as the only
difference is what chunk class to load.

Also added a ton of debug messages in trying to pin down this save bug;
I'll take those out as needed.

Next: auto-update the maps for saved regions, as I didn't even realize it
was working until I manually regenerated the maps.